ref(docs): rewrite with mkdocs
The docs is written in reStructuredText using Sphinx, which is a tool to easily make beautiful
documentation. This has been a great tool for writing our documentation as it included features
like auto-docs for our controller API - which is written in Python - as well as custom themes on
ReadTheDocs. Most of the early contributors were comfortable with ReST when working on Deis v1.
Most of the code was written in Python, so Sphinx was the best choice at the time. Enter 2015, and
Deis v2 is now mostly written in Go. Save for Godeps/, Deis's language distribution per-file is 45%
Python, 40% Go, and 15% other (documentation, bash scripts, YAML/JSON/TOML, Makefiles, Dockerfiles,
etc). This does not include projects like helm/helm and deis/pkg. In short, our usage of Go is
ever-growing and we are drawing a newer crowd to the project. This crowd is comfortable
contributing to Sphinx, but newer contributors are more comfortable writing documentation in more
modern markup languages like Markdown.

We have experimented with using MkDocs with our newer projects like Helm and have seen success in
both readability and contribution efforts, so now is a good opportunity to re-write the Deis
documentation in Markdown while half of it needs to be re-written in preparation for v2. The
auto-docs is a handy feature, but only the core contributors have been taking advantage of the
feature so it won't be missed all that much in favor of greater contributor adoption and ease of
use by tools like MkDocs.

NOTE: some documentation is completely missing pending how we are moving forward on those sections
for v2. This is mostly to do with the "Managing Deis" side of things, since logging, monitoring,
etc. are potentially all stale for v2.

To test this:

```
$ pip install -r docs/requirements.txt
$ mkdocs serve
```