fix(settings): default to empty host environment variable
If you do not supply a host variable, Then psycopg2 will connect using a Unix socket in the same
manner as psql. Surprisingly, when you specify a HOST, psycopg2 will connect with TCP/IP, which
requires a password. This broke my environment with the following:

        OperationalError: fe_sendauth: no password supplied

See: http://stackoverflow.com/a/23871618