fix(scheduler): use app name when creating services
When we eventually allow applications to be grouped together in the same namespace, we'll want to
have the service names discoverable to other instances in the same namespace. Having a name like
"app-1234" was created so we could get over the limitation of 24 characters for service names since
applications in v1 could be as long as 64 characters in length. Now that we're working on v2, we
can just restrict app names to 24 characters.

The app name validation regex was not changed because the docker repository name regex is a subset
of the kubernetes service name regex, so we still need to keep ourselves restricted to that name
regex so we can still push those images to the internal registry.