fix(certificates): only delete k8s secrets for certs when last domain associated is detached
Before the code would delete the secret on every detach but now instead it checks if there are any domains associated still with the certificate before deleting.
It can do so before the association between domain and cert is nulled out a few lines up.

This wasn't caught since the exception rule is to log the error instead of blowing up. There are tests that expose this behaviour but only against a real k8s cluster