ref(apps): iterate over RC for desired pod count when doing an app restart
Rationale here is that the RC already has the desired information and we may be iterating over many RCs.
Not using the pod count is helpful in case other operations happen to have interfered with things.

Another option is to use the structure info from the App model but this solutions seems more solid