fix(api): pass if no builds are associated with the release
If an application only has releases with no builds (such as a `deis create`),
the restoration process will bail out because an exception was raised,
killing the controller. However, this is not a concern if no build
is associated with the release, so we can just check and ignore
this case.