ref(port): when using a private registry then look for a PORT env var instead of auto discovering port
Rationale here is to not pull down the image since Kubernetes now owns the downloading and interaction with the Private Registry. Pulling an image down into the Deis Registry will strip away the benefits of the security given.

This change moves the docker port auto-discovery into the docker-client and general port handling into Release model, making the Scheduler unaware of the inherent logic. Ports are still auto-discovered on public images but when Private Registry is used the default port is 5000 and if the user wants to use a different one then they need to do config:set PORT=3000

Logic may change in the future where a user can use Private Registry but pull into the local registry, and get port auto-discovery, at the cost of security.