fix(config): when rolling a new config copy from the latest release instead of latest config for the app
This helps mitigate #798 for day to day use (quick fix) as the previous config is no longer pulled from the top of the application config stack and instead pulls the latest release config.
Doing that will prevent the logic from re-using a potentially broken Config object. When a Release object is deleted an Config object can be orphaned but still available.