feat(*): add Deployments support behind a feature flag
This adds Deployment object support (replaces ReplicationController) when a feature flag is enabled. This can be done by either settings a global ENV Var on the Controller or setting a config per app. The variable is DEIS_KUBERNETES_DEPLOYMENT=1

A few new features (Deployments only) have been introduced or specifically expose Deployment functionalities:
* KUBERNETES_DEPLOYMENTS_REVISION_HISTORY_LIMIT (global and per app setting) specifies how many revisions (ReplicaSets) to keep
* Stop concurrent deployments (triggered by push, pull, config:set and so on)

If Deployments is enabled on an existing app then a whole new copy will be spun up (behind the scenes) as Deployment and then the old ReplicationController is spun down. This is a one way conversion.

Note that pod naming has changed due to how Deployments adds a hash to the name, among other things.