fix(scheduler): allow Deployment "in progress" to be bypassed in case of errors or timeout (#978)
It is possible to get a Deployment into a state where it can not bring up pods. The API will think that means that the Deployment is still "in progress". The scenario around this mostly happens when the client connection (CLI) is lost and the API loses context.

This change brings in the ability to get past the "in progress" stop sign when the Deployment is found to have errors (image not found, limits are bad, etc) or it has been in that state for more than the overall deploy timout limit is