tests(scheduler): confirm "network unreachable" raises KubeHTTPException
This error occurs when the registry hosting the image being pulled by `deis pull` (or, in the case
of `git push deis master`, quay.io/deis/slugrunner) becomes unreachable. This test confirms that
master has fixed the original error, which is that a KubeException used to be raised, being
uncaught in older releases of Workflow. This has since been fixed in master.