fix(boot): Don't change group ownership of docker socket
This is a better fix for issue #804. Instead of changing the gid of the
docker socket to that of the deis group in the container, which affects
access to the docker socket by existing users on the host, we now either
add the deis user to an existing group in the container with the same
gid or create a new "docker" group with the same gid as the socket if no
existing group has that gid.